backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-multipart==0.0.6
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
scipy==1.11.4
openai==1.3.0
requests==2.31.0
psycopg2-binary==2.9.9
asyncpg==0.29.0